// src/app/api/model/vehicle.ts
/**
 * Leasing Service
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * Contact: lease-service@allane.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Vehicle {
    id?: number;
    brand: string;
    model: string;
    modelYear: number;
    /**
     * Vehicle identification number when the vehicle is produced. During creation of contract it could be empty.
     */
    vin?: string;
    price: number;
}